3rd commit

+++ client.c
@@ -26,18 +26,60 @@ int	ft_atoi(const char *str)
 	return (nbr * sign);
 }
 
+void	decimal_to_bits(unsigned char c, unsigned char *strbit, int msgsize)
+{
+	int	i;
+	static int	j;
+
+	i = 7;
+	//if (j != 0)
+	//	strbit[j++] = ' ';
+	while (i >= 0)
+	{
+		strbit[j] = (c >> i & 1) + '0';
+		i--;
+		j++;
+	}
+	if (j == msgsize * 8)
+		strbit[j] = '\0';
+	ft_printf("j : %d c : %c (%d)\n", j, c, c);
+}
+
 int main(int    argc, char  *argv[])
 {
     pid_t   serverpid;
-	char	*message;
 	int	msgsize;
+	unsigned char	*strbit;
+	int	i;
 
-    serverpid = ft_atoi(argv[1]);
-	message = argv[2];
+	serverpid = ft_atoi(argv[1]);
 	msgsize = ft_strlen(argv[2]);
-	while (msgsize--)
-    	kill(serverpid, SIGUSR1);
-    while (1)
-        continue ;
+	strbit = malloc(sizeof(unsigned char) * msgsize * 10);
+	i = 0;
+	while (i < msgsize)
+	{
+		decimal_to_bits(argv[2][i], strbit, msgsize);
+		i++;
+		ft_printf("%s\n", strbit);
+	}
+	i = 0;
+	while (strbit[i] != '\0')
+		i++;
+	ft_printf("str size : %d\n", i);
+	i = 0;
+	while (strbit[i] != '\0')
+	{
+		if (strbit[i] == '0')
+		{
+			kill(serverpid, SIGUSR1);
+			ft_printf("0");
+		}
+		if (strbit[i] == '1')
+		{
+			kill(serverpid, SIGUSR2);
+			ft_printf("1");
+		}
+		i++;
+	}
     return (0);
 }
